organize EmptyCart on its own folder
