+++ src/components/cart/EmptyCart/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './EmptyCart'
